aider: Ensure Poetry binary path is correctly set for the Fish shell environment.
# Aider chat conversation:

USER: I don't use bash, I use fish.

ASSISTANT: Understood. Since you are using the Fish shell, I will make the necessary changes to the Dockerfile to ensure that the `PATH` is correctly set for the Fish shell environment. This involves appending the Poetry binary path to the `fish_user_paths` universal variable, which is the equivalent of adding to the `PATH` in Fish.

Here are the diffs for those changes:

```diff
--- .devcontainer/Dockerfile
+++ .devcontainer/Dockerfile
@@ ... @@
 # Install Poetry
 RUN curl -sSL https://install.python-poetry.org | python3 - && \
-    echo 'export PATH="/root/.local/bin:$PATH"' >> $HOME/.bashrc && \
+    mkdir -p /home/vscode/.config/fish/conf.d && \
+    echo 'set -U fish_user_paths /root/.local/bin $fish_user_paths' >> /home/vscode/.config/fish/conf.d/poetry_path.fish && \
     /root/.local/bin/poetry config virtualenvs.create false
@@ ... @@
 # Use the installed Poetry to set up the environment
-ENV PATH="/root/.local/bin:$PATH"
 RUN poetry env use python3.11
 RUN poetry install --no-interaction --no-root -n $(nproc)
-ENV PATH="/workspaces/${localWorkspaceFolderBasename}/.venv/bin:$PATH"
+RUN echo 'set -U fish_user_paths /workspaces/${localWorkspaceFolderBasename}/.venv/bin $fish_user_paths' >> /home/vscode/.config/fish/config.fish
```

These changes will ensure that the Poetry binary path is added to the `fish_user_paths` universal variable for the `vscode` user, which is the user that will be running the shell in the devcontainer. This will make the `poetry` command available in the Fish shell environment.